requirements.txt
Flask==3.1.1
langchain==0.3.26
langchain-community==0.3.26
langchain-openai==0.3.24
langchain-pinecone==0.2.8
sentence-transformers==4.1.0
pypdf==5.6.1
python-dotenv==1.1.0
pinecone
flask-cors==5.0.0
langchain-huggingface
langchain-core